+++ bindings/src/vocab/mod.rs
@@ -29,3 +29,9 @@ pub fn prune_universe(data: &str, universe: &str, min_count: Option<u32>, output
 
     Ok(())
 }
+
+#[pymodule]
+pub fn vocab(_py: Python, m: &PyModule) -> PyResult<()> {
+    m.add_wrapped(wrap_pyfunction!(prune_universe))?;
+    Ok(())
+}